pyproject.toml
[tool.poetry]
name = "dench"
version = "0.1.0"
description = ""
authors = ["Rafayel Gasparyan <rafayelgasparyan04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
pillow = "^10.3.0"
dj-static = "^0.0.6"
django-cleanup = "^8.1.0"
python-environ = "^0.4.54"
django-ckeditor = "^6.7.1"
requests = "^2.32.3"
polygonscan-python = "^1.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
